pyproject.toml
[tool.poetry]
name = "mailing-app"
version = "0.1.0"
description = ""
authors = ["reloverevolt <savitskiy.revenway@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "3.2.12"
djangorestframework = "3.13.1"
APScheduler = "^3.9.1"
django-extensions = "^3.2.0"
django-dramatiq = "^0.11.0"
pika = "^1.3.0"
requests = "^2.28.1"
gunicorn = "^20.1.0"
flake8 = "^5.0.4"
black = "^22.6.0"
psycopg2-binary = "^2.9.3"
pytz = "^2022.2.1"
PyYAML = "^6.0"
marshmallow-dataclass = "^8.5.8"
pytz-deprecation-shim = "^0.1.0"
tzdata = "^2022.2"
drf-yasg = "^1.21.3"
pre-commit = "^2.20.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
